Changed some uint to size_t. Added switch to disable OpenMP in compilation
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@1815 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ BayesTest.cpp
@@ -4,7 +4,7 @@
 #include <cstdlib>
 #include "BayesTest.h"
 
-BayesTest::BayesTest(unsigned int aNumSites)
+BayesTest::BayesTest(size_t aNumSites)
 {
 	mNumSites = aNumSites;
 	mSiteClassProb = new double[4*mNumSites];
